#33 [fix] erd 수정
