--- src/main/java/com/example/letscareer/always/dto/request/AlwaysRequest.java
@@ -1,4 +1,0 @@
-package com.example.letscareer.always.dto.request;
-
-public record AlwaysRequest() {
-}
