+++ src/main/java/com/example/letscareer/mid_review/service/MidReviewService.java
@@ -61,82 +61,6 @@ public class MidReviewService {
 
         midReviewRepository.save(midReview);
     }
-
-    public FastReviewsResponse getFastReviews(Long userId, int page, int size) {
-
-        LocalDate today = LocalDate.now();
-        LocalDate threeDaysLater = today.plusDays(4); //3일 뒤까지 포함
-
-        Pageable pageable = PageRequest.of(page - 1, size); // JPA는 페이지 인덱스가 0부터 시작
-
-        // 면접 회고 조회
-        Page<IntReview> intReviews = intReviewRepository.findAllByUserIdAndStageDeadlineWithin3Days(userId, today, threeDaysLater, pageable);
-
-        // 중간 회고 조회
-        Page<MidReview> midReviews = midReviewRepository.findAllByUserIdAndStageDeadlineWithin3Days(userId, today, threeDaysLater,  pageable);
-
-        List<FastDTO> fastReviews = new ArrayList<>();
-
-        // 면접 회고 데이터를 응답 형식에 맞게 변환
-        intReviews.forEach(ir -> {
-            if (ir.getStage().getAlways() == null) {
-                FastDTO fastDTO = new FastDTO(
-                        ir.getStage().getStageId(),
-                        ir.getIntReviewId(),
-                        ir.getStage().getSchedule().getCompany(),
-                        ir.getStage().getSchedule().getCompany(),
-                        "면접",
-                        calculatePlusDays(ir.getStage().getDate())
-                );
-                fastReviews.add(fastDTO);
-            }else{
-                FastDTO fastDTO = new FastDTO(
-                        ir.getStage().getStageId(),
-                        ir.getIntReviewId(),
-                        ir.getStage().getAlways().getCompany(),
-                        ir.getStage().getAlways().getCompany(),
-                        "면접",
-                        calculatePlusDays(ir.getStage().getDate())
-                );
-                fastReviews.add(fastDTO);
-            }
-        });
-
-        // 중간 회고 데이터를 응답 형식에 맞게 변환
-        midReviews.forEach(mr -> {
-            if (mr.getStage().getAlways() == null) {
-                FastDTO fastDTO = new FastDTO(
-                        mr.getStage().getStageId(),
-                        mr.getMidReviewId(),
-                        mr.getStage().getSchedule().getCompany(),
-                        mr.getStage().getSchedule().getCompany(),
-                        "중간",
-                        calculatePlusDays(mr.getStage().getDate())
-                );
-                fastReviews.add(fastDTO);
-            }else{
-                FastDTO fastDTO = new FastDTO(
-                        mr.getStage().getStageId(),
-                        mr.getMidReviewId(),
-                        mr.getStage().getAlways().getCompany(),
-                        mr.getStage().getAlways().getCompany(),
-                        "중간",
-                        calculatePlusDays(mr.getStage().getDate())
-                );
-                fastReviews.add(fastDTO);
-            }
-        });
-
-        // plusDay 값으로 fastReviews 리스트 정렬
-        fastReviews.sort(Comparator.comparingInt(FastDTO::plusDay));
-
-        return new FastReviewsResponse(
-                page,
-                size,
-                fastReviews.size(),
-                fastReviews
-        );
-    }
     public static int calculatePlusDays(LocalDate date) {
         LocalDate now = LocalDate.now();
