+++ src/main/java/com/example/letscareer/stage/repository/StageRepository.java
@@ -22,5 +22,5 @@ public interface StageRepository extends JpaRepository<Stage, Long> {
 
     @Query("SELECT st FROM Stage st WHERE st.schedule.user.userId = :userId AND st.date = :date")
     List<Stage>findAllByUserIdAndDate(@Param("userId") Long userId, @Param("date") LocalDate date);
-    Optional<Stage> findTopByAlwaysAlwaysIdOrderByOrderDesc(Long alwaysId);
+    Optional<Stage> findTopByScheduleScheduleIdOrderByOrderDesc(Long scheudleId);
 }
